[PORT] Port stock_split_picking to 10.0
Make some changes to the basic workflow, as the picking workflow was
changed in 9.0. There is no longer a separate wizard for processing
the packing operations, rather they are visible on the picking
form. As such, we emulate the normal backorder creation as
closely as possible, while not actually creating a backorder.